fix: comprehensive testing — fix 4 critical bugs, remove dead code, update tests
Critical bugs fixed:
- Export functions destroyed by wrong aliases in goobi_xml.py (8/10 test failures)
- Dashboard thumbnail path mismatch (/api/images/{id} vs /api/images/{id}/data)
- XSS vulnerability in renderImgGrid template literal
- Image cleanup only cleared memory, not disk files

Architecture cleanup:
- Remove dead monolithic app.py (783 lines), rename app_new.py to app.py
- Remove orphaned ner_hybrid_fix.py
- Clean up duplicate imports in routes/ai.py

Test fixes:
- Security tests now check deps.py/routes/ instead of dead app.py
- Export tests now set up required field mapping
- All 369 tests pass (was: 5 failures + 5 errors)

Added comprehensive test report: docs/TESTBERICHT.md
