+++ src/kwb/export/goobi_xml.py
@@ -412,6 +412,3 @@ def dataframe_to_goobi_xml_files(
         written.append(path)
 
     return written
-# Backward compatibility for API imports
-export_goobi_xml = dataframe_to_goobi_xml
-export_goobi_batch = dataframe_to_goobi_xml_files
